refactor: :recycle: move address from channel to provider
