+++ apps/taxes/src/modules/app/get-app-config.test.ts
@@ -20,6 +20,13 @@ const mockedProviders: ProvidersConfig = [
         password: "avatax-password",
         username: "avatax-username",
       },
+      address: {
+        city: "New York",
+        country: "US",
+        state: "NY",
+        street: "123 Main St",
+        zip: "10001",
+      },
     },
   },
   {
@@ -31,6 +38,13 @@ const mockedProviders: ProvidersConfig = [
       credentials: {
         apiKey: "taxjar-api-key",
       },
+      address: {
+        city: "New York",
+        country: "US",
+        state: "NY",
+        street: "123 Main St",
+        zip: "10001",
+      },
     },
   },
 ];
@@ -38,13 +52,6 @@ const mockedEncryptedProviders = encrypt(JSON.stringify(mockedProviders), mocked
 
 const mockedChannels: ChannelsConfig = {
   "default-channel": {
-    address: {
-      city: "New York",
-      country: "US",
-      state: "NY",
-      street: "123 Main St",
-      zip: "10001",
-    },
     providerInstanceId: "1",
   },
 };
