+++ apps/taxes/src/modules/channels-configuration/channels-config.ts
@@ -1,19 +1,8 @@
 import { z } from "zod";
 import { ChannelFragment } from "../../../generated/graphql";
 
-const addressSchema = z.object({
-  country: z.string(),
-  zip: z.string(),
-  state: z.string(),
-  city: z.string(),
-  street: z.string(),
-});
-
-export type ChannelAddress = z.infer<typeof addressSchema>;
-
 export const channelSchema = z.object({
   providerInstanceId: z.string(),
-  address: addressSchema,
 });
 export type ChannelConfig = z.infer<typeof channelSchema>;
 
@@ -22,13 +11,6 @@ export type ChannelsConfig = z.infer<typeof channelsSchema>;
 
 export const defaultChannelConfig: ChannelConfig = {
   providerInstanceId: "",
-  address: {
-    city: "",
-    country: "",
-    state: "",
-    street: "",
-    zip: "",
-  },
 };
 
 export const createDefaultChannelsConfig = (channels: ChannelFragment[]): ChannelsConfig => {
